+++ src/init.c
@@ -4,9 +4,11 @@
 #include <Rinternals.h>
 
 extern SEXP flounder_rust_capabilities(void);
+extern SEXP flounder_pod5_find(SEXP path);
 
 static const R_CallMethodDef CallEntries[] = {
     {"flounder_rust_capabilities", (DL_FUNC) &flounder_rust_capabilities, 0},
+    {"flounder_pod5_find", (DL_FUNC) &flounder_pod5_find, 1},
     {NULL, NULL, 0}
 };